+++ docs/ci.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+set -ex
+
+(
+    cd docker \
+    && docker compose build \
+    && docker compose run --rm sphinx
+)